# openapi/libraryz.yaml
openapi: 3.1.0
info:
  title: LibraryZ API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: |
    LibraryZ backend: auth and sessions, catalog (works + editions with
    content-addressed file storage), crowdsourced edits with moderation,
    personal libraries, and recommendations. See PLAN.md and ARCHITECTURE.md
    in the repo root.

    **Every operation has a cURL sample** (`x-codeSamples`). They assume:

    ```sh
    BASE=http://localhost:8080          # Kong in `docker compose up`
    TOKEN=...      REFRESH=...          # from POST /auth/login (see its sample)
    WORK_ID=...    EDITION_ID=...    CONTRIBUTION_ID=...
    ```

    **Authentication.** `POST /auth/login` returns a short-lived access token
    (JWT, 15 minutes by default) and a single-use refresh token. Send the
    access token as `Authorization: Bearer <token>`. On a `401`, call
    `POST /auth/refresh` once and retry; if that also fails, log in again.
    Presenting an already-used refresh token revokes the whole session.

    **Errors.** `4xx`/`5xx` bodies are short `text/plain` messages (`4xx`
    messages are safe to show to users), except where an operation documents
    a JSON body (e.g. the `409` on edition upload). `5xx` bodies are generic;
    details are only logged server-side.

    **Rate limits.** In the compose/production setup Kong enforces per-IP
    limits (login 5/min, signup 3/min, refresh/logout 20/min, edition upload
...
servers:
...
    description: Local dev
...
security: []
paths:
...
  /works:
    get:
      operationId: listWorks
      tags: [catalog]
      summary: List works
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/works?limit=50&offset=0"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Work" }
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      operationId: createWork
      tags: [catalog]
      summary: Create a new work
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkRequest"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X POST "$BASE/works" -H "Authorization: Bearer $TOKEN" -H 'Content-Type: application/json' \
              -d '{"title":"Moby-Dick","authors":"Herman Melville","publication_year":1851}'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Work" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /works/search:
    get:
      operationId: searchWorks
      tags: [catalog]
      summary: Search works by free-text query
      description: |
        Matches against `title`, `subtitle`, `authors`, and `description`,
        most relevant first. On Postgres uses a weighted `tsvector` GIN index
        (`simple` configuration — no stemming / stopwords) queried with
        `websearch_to_tsquery`, so `"exact phrase"`, `or`, and `-exclude`
        work; results are ordered by `ts_rank` (title > subtitle/authors >
        description), then `created_at DESC`. On sqlite (the test DB) falls
        back to a literal, case-insensitive substring match (`%` and `_` are
        not wildcards) ordered title hits, then subtitle/author hits, then
        description hits, then `created_at DESC`. Empty `q` returns 400.
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -G "$BASE/works/search" \
              --data-urlencode 'q="herman melville" -abridged' \
              --data-urlencode limit=20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Work" }
        "400":
          $ref: "#/components/responses/BadRequest"
  /works/{id}:
    get:
      operationId: getWork
      tags: [catalog]
      summary: Get a single work with its editions and tags
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/works/$WORK_ID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Work" }
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
    delete:
      operationId: removeWork
      tags: [moderation]
      summary: Remove a work (moderator, or its creator while empty)
      description: |
        Moderators can take down any work; any other user can remove only a
        work they created that has no editions (the upload client does this
        when a new work's first upload is rejected). Removes the work and all
        of its editions. Soft
        delete: the rows are kept with who/when/why for audit (and can be
        restored in the DB), but the work disappears from list, search, get,
        download, personal libraries and recommendations. The files are
        purged by the blob GC after `LIBRARYZ_BLOB_GC_RETENTION`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RemovalRequest" }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X DELETE "$BASE/works/$WORK_ID" -H "Authorization: Bearer $TOKEN" -H 'Content-Type: application/json' \
              -d '{"reason":"DMCA notice #42"}'
      responses:
        "204":
          description: Removed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /works/{id}/editions:
    post:
      operationId: uploadEdition
      tags: [editions]
      summary: Upload a new edition (file) for a work
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [format, file]
              properties:
                format:
                  type: string
                  description: "pdf | epub | mobi | azw3 | djvu | cbz | txt | ..."
                language:
                  type: string
                  description: BCP-47 language tag (optional)
                file:
                  type: string
                  format: binary
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X POST "$BASE/works/$WORK_ID/editions" -H "Authorization: Bearer $TOKEN" \
              -F format=epub -F language=en -F file=@moby-dick.epub
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Edition" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            The file is already in the library as an edition, on this work or
            another one. Matched on the sha256 of the uploaded bytes, or of the
            bytes after sanitization. Nothing is recorded. If that edition was
            removed by a moderator, `edition_id` / `work_id` are omitted and
            `error` says so (a takedown can't be undone by re-uploading).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DuplicateEdition" }
        "413":
          description: >
            The file couldn't be sanitized within the server's memory or time
            budget (for example, a PDF whose compressed object streams
            inflate to gigabytes). Nothing is recorded.
          content:
            text/plain:
              schema: { type: string }
        "415":
          description: Format is not on the allowlist (currently PDF, EPUB, TXT)
          content:
            text/plain:
              schema: { type: string }
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /editions/{id}:
    get:
      operationId: getEdition
      tags: [editions]
      summary: Get edition metadata
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/editions/$EDITION_ID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Edition" }
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
    delete:
      operationId: removeEdition
      tags: [moderation]
      summary: Remove an edition (moderator)
      description: |
        Moderator-only takedown of one edition (soft delete; see
        `DELETE /works/{id}`). Re-uploading byte-identical content afterwards
        is refused with 409.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RemovalRequest" }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X DELETE "$BASE/editions/$EDITION_ID" -H "Authorization: Bearer $TOKEN" -H 'Content-Type: application/json' \
              -d '{"reason":"broken file"}'
      responses:
        "204":
          description: Removed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /editions/{id}/download:
    get:
      operationId: downloadEdition
      tags: [editions]
      summary: Download the edition bytes
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            # -OJ saves under the server-provided "<Title> - <Authors>.<format>" name.
            curl -s -OJ "$BASE/editions/$EDITION_ID/download"
      responses:
        "200":
          description: File stream
          headers:
            Content-Disposition:
              schema: { type: string }
            X-Content-SHA256:
              schema: { type: string, description: "Hex sha256 of the bytes (matches Edition.sha256)" }
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token from `POST /auth/login` or `POST /auth/refresh`. Expires
        after `LIBRARYZ_ACCESS_TOKEN_TTL` (15m); revoked immediately by
        `POST /auth/logout-all`.
  parameters:
    UUIDPathId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Bad request
      content:
        text/plain:
          schema: { type: string }
    Unauthorized:
      description: >
        Missing, malformed, expired or revoked access token (refresh it with
        `POST /auth/refresh`), or wrong credentials on login
      content:
        text/plain:
          schema: { type: string }
    Forbidden:
      description: Authenticated but not permitted (e.g. moderator required)
      content:
        text/plain:
          schema: { type: string }
    NotFound:
      description: Resource not found
      content:
        text/plain:
          schema: { type: string }
    TooManyRequests:
      description: Rate limit exceeded (enforced by the Kong gateway, per client IP)
      headers:
        Retry-After:
          description: Seconds until the limit resets
          schema: { type: integer }
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string, example: API rate limit exceeded }
...
  schemas:
...
    RemovalRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 1000
          description: Why it was taken down — kept for the audit trail.
...
    DuplicateEdition:
      type: object
      required: [error]
      properties:
        error:      { type: string, example: this file is already in the library }
        edition_id: { type: string, format: uuid, description: "The existing edition (absent if it was removed by a moderator)" }
        work_id:    { type: string, format: uuid, description: The work the existing edition belongs to }
    CreateWorkRequest:
      type: object
      description: >
        Only these fields are read. Server-owned fields sent in the body
        (id, created_at, updated_at, editions, tags) are ignored; editions
        are only created by uploading a file.
      required: [title]
      properties:
        title:            { type: string }
        subtitle:         { type: string }
        authors:          { type: string, description: "Phase 1: semicolon-separated. Normalized in Phase 2." }
        description:      { type: string }
        language:         { type: string }
        publication_year: { type: integer }
        isbn:             { type: string }
        openlibrary_id:   { type: string }
    Work:
      type: object
      properties:
        id:               { type: string, format: uuid }
        title:            { type: string }
        subtitle:         { type: string }
        authors:          { type: string }
        description:      { type: string }
        language:         { type: string }
        publication_year: { type: integer }
        isbn:             { type: string }
        openlibrary_id:   { type: string }
        created_at:       { type: string, format: date-time }
        updated_at:       { type: string, format: date-time }
        editions:
          type: array
          items: { $ref: "#/components/schemas/Edition" }
        tags:
          type: array
          items: { $ref: "#/components/schemas/Tag" }
    Edition:
      type: object
      properties:
        id:          { type: string, format: uuid }
        work_id:     { type: string, format: uuid }
        format:      { type: string }
        language:    { type: string }
        size_bytes:  { type: integer, format: int64 }
        sha256:      { type: string }
        uploaded_by: { type: integer }
        created_at:  { type: string, format: date-time }
    Tag:
      type: object
      properties:
        id:   { type: string, format: uuid }
        name: { type: string }
...